+++ tests/test_docker.py
@@ -1103,7 +1103,7 @@ class ImageTestCase(unittest.TestCase):
                 ),
                 expected_name='image',
                 expected_tag='latest',
-                expected_registry='',
+                expected_registry=None,
                 expected_str='image:latest',
             ),
             with_tag=dict(
@@ -1113,7 +1113,7 @@ class ImageTestCase(unittest.TestCase):
                 ),
                 expected_name='image',
                 expected_tag='tag',
-                expected_registry='',
+                expected_registry=None,
                 expected_str='image:tag',
             ),
             with_registry=dict(
@@ -1154,7 +1154,7 @@ class ImageTestCase(unittest.TestCase):
                 ),
                 expected_name='image',
                 expected_tag='tag',
-                expected_registry='',
+                expected_registry=None,
                 expected_str='image:tag',
             ),
             single_arg_with_registry=dict(
@@ -1182,7 +1182,7 @@ class ImageTestCase(unittest.TestCase):
                 ),
                 expected_name='image',
                 expected_tag='foo',
-                expected_registry='',
+                expected_registry=None,
                 expected_str='image:foo',
             ),
             forced_with_registry=dict(
@@ -1324,13 +1324,13 @@ class ImageTestCase(unittest.TestCase):
             name=dict(
                 image='image',
                 expected_name='image',
-                expected_registry='',
+                expected_registry=None,
                 expected_tag='latest',
             ),
             name_and_tag=dict(
                 image='image:tag',
                 expected_name='image',
-                expected_registry='',
+                expected_registry=None,
                 expected_tag='tag',
             ),
             name_and_registry=dict(